\name{RcmdrSurvivalTPlugin-internal}
\title{Internal RcmdrSurvivalTPlugin objects}
\alias{fncKaplanMeier}
\alias{fncLogRankTest}
\alias{fncLogLogSurvival}
\alias{fncCoxModel}
\alias{fncCoxInfluencePlot}
\alias{fncCoxPHSchoenfeld}
\alias{fncCoxPHSchoenfeldGraph}
\alias{simulateConfidenceIntervals}
\alias{listCoxModels}
\alias{CoxP}
\alias{formulaFieldsCox}
\alias{modelFormulaCox}
\description{Internal RcmdrSurvivalTPlugin objects.}
\details{These are not to be called by the user.}
\keyword{internal}
