\name{RcmdrSurvivalTPlugin-package}
\alias{RcmdrSurvivalTPlugin-package}
\alias{RcmdrSurvivalTPlugin}
\docType{package}
\title{
Install the SurvivalT Rcmdr Plug-In
}
\description{
This package provides an Rcmdr \dQuote{plug-in} based on the survival package for easier student access to survival analysis.
}
\details{
\tabular{ll}{
Package: \tab RcmdrSurvivalTPlugin\cr
Type: \tab Package\cr
Version: \tab 1.0-7\cr
Date: \tab 2008-11-18\cr
License: \tab GPL version 2 or newer\cr
}
}

\author{
Daniel Leucuta <danny.ldc@gmail.com>

Maintainer: Daniel Leucuta <danny.ldc@gmail.com>
}

\keyword{ package }
\seealso{
\code{\link[Rcmdr]{Rcmdr}}.
}

